+++ src/Billing/Billing.php
@@ -2,6 +2,7 @@
 
 namespace Hyvor\Internal\Billing;
 
+use Hyvor\Internal\Billing\Dto\LicenseOf;
 use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\Exception\LicenseOfCombinationNotFoundException;
 use Hyvor\Internal\Billing\License\License;
@@ -9,10 +10,19 @@ use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\Component\InstanceUrlResolver;
 use Hyvor\Internal\InternalApi\Exceptions\InternalApiCallFailedException;
 use Hyvor\Internal\InternalApi\InternalApi;
-use Hyvor\Internal\InternalApi\InternalApiMethod;
+use Hyvor\Internal\InternalConfig;
+use Hyvor\Internal\Util\Crypt\Encryption;
 
 class Billing
 {
+
+    public function __construct(
+        private InternalConfig $internalConfig,
+        private InstanceUrlResolver $instanceUrlResolver,
+        private InternalApi $internalApi,
+        private Encryption $encryption
+    ) {
+    }
 
     /**
      * @return array{token: string, urlNew: string, urlChange: string}
@@ -24,7 +34,7 @@ class Billing
         bool $isAnnual,
         ?Component $component = null,
     ): array {
-        $component ??= Component::current();
+        $component ??= $this->internalConfig->getComponent();
 
         // this validates the plan name as well
         $plan = $component->plans()->getPlan($planName);
@@ -38,9 +48,9 @@ class Billing
             $isAnnual,
         );
 
-        $token = $object->encrypt();
+        $token = $this->encryption->encrypt($object);
 
-        $baseUrl = InstanceUrlResolver::getInstanceUrl() . '/account/billing/subscription?token=' . $token;
+        $baseUrl = $this->instanceUrlResolver->publicUrlOfCore() . '/account/billing/subscription?token=' . $token;
 
         return [
             'token' => $token,
@@ -69,14 +79,13 @@ class Billing
      */
     public function licenses(array $of, ?Component $component = null): LicensesCollection
     {
-        $component ??= Component::current();
+        $component ??= $this->internalConfig->getComponent();
 
         /**
          * @var array{user_id: int, resource_id: ?int, license: ?string}[] $response
          */
-        $response = InternalApi::call(
+        $response = $this->internalApi->call(
             Component::CORE,
-            InternalApiMethod::GET,
             '/billing/licenses',
             [
                 'of' => array_map(fn($ofOne) => $ofOne->toArray(), $of)