+++ src/Billing/SubscriptionIntent.php
@@ -4,11 +4,13 @@ namespace Hyvor\Internal\Billing;
 
 use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\Util\Transfer\Encryptable;
+use Symfony\Component\DependencyInjection\Attribute\Exclude;
 
 /**
  * This is the object for requesting a new subscription
  * or a plan change.
  */
+#[Exclude]
 class SubscriptionIntent
 {