+++ tests/Unit/Billing/BillingSymfonyTest.php
@@ -7,7 +7,7 @@ use Hyvor\Internal\Billing\Dto\LicenseOf;
 use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\SubscriptionIntent;
 use Hyvor\Internal\InternalApi\InternalApi;
-use Hyvor\Internal\Tests\LaravelTestCase;
+use Hyvor\Internal\Tests\SymfonyTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
@@ -17,23 +17,29 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 #[CoversClass(SubscriptionIntent::class)]
 #[CoversClass(LicensesCollection::class)]
 #[CoversClass(LicenseOf::class)]
-class BillingLaravelTest extends LaravelTestCase
+class BillingSymfonyTest extends SymfonyTestCase
 {
 
     use BillingTestTrait;
 
     protected function getBilling(): Billing
     {
-        return app(Billing::class);
+        $billing = $this->container->get(Billing::class);
+        assert($billing instanceof Billing);
+        return $billing;
     }
 
     protected function setHttpResponse(MockResponse $response): void
     {
-        app()->singleton(HttpClientInterface::class, fn() => new MockHttpClient($response));
+        $currentClient = $this->container->get(HttpClientInterface::class);
+        assert($currentClient instanceof MockHttpClient);
+        $currentClient->setResponseFactory($response);
     }
 
     protected function getInternalApi(): InternalApi
     {
-        return app(InternalApi::class);
+        $internalApi = $this->container->get(InternalApi::class);
+        assert($internalApi instanceof InternalApi);
+        return $internalApi;
     }
 }