+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 - InstanceUrl removed, use InternalConfig and ComponentUrl instead
 - InternalApi\ComponentType moved to Component\Component, and many static methods have been removed
+- Component::current() is removed, use InternalConfig instead
 - Auth methods are no longer static
 - Auth login, signup, and logout methods now return a Symfony RedirectRepsonse object
 - component label is removed from metrics, add that in Prometheus instead
+- InternalApiMethod is removed, call internal API requests should be POST now
+- PHP 8.4 required
 
 ## 2.1.0 - 2025-03-23